Fixes module import issue
- Fixes circular dependency in module code
