+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cloudnativetoolkit/iascable",
-  "version": "3.0.0-beta.4",
+  "version": "3.0.0-beta.5",
   "description": "",
   "keywords": [],
   "main": "dist/iascable.umd.js",
@@ -116,12 +116,14 @@
     "@types/inquirer": "^7.3.1",
     "@types/jest": "^26.0.20",
     "@types/js-yaml": "^4.0.0",
+    "@types/lodash.camelcase": "^4.3.7",
     "@types/lodash.clonedeep": "^4.5.6",
     "@types/lodash.unionby": "^4.8.7",
     "@types/lodash.uniq": "^4.5.7",
     "@types/lodash.uniqby": "^4.7.7",
     "@types/lodash.uniqwith": "^4.5.7",
     "@types/node": "^14.14.27",
+    "@types/rollup-plugin-json": "^3.0.3",
     "@types/superagent": "^4.1.10",
     "colors": "^1.3.2",
     "commitizen": "^4.2.4",
